+++ lib/data-models/account.dart
@@ -9,6 +9,7 @@ import 'package:vocdoni/lib/model-base.dart';
 import 'package:eventual/eventual.dart';
 import 'package:vocdoni/data-models/entity.dart';
 import 'package:vocdoni/lib/globals.dart';
+import 'package:vocdoni/lib/random.dart';
 
 /// This class should be used exclusively as a global singleton.
 /// AccountPoolModel tracks all the registered accounts and provides individual models that
@@ -477,6 +478,13 @@ class AccountModel implements ModelRefreshable, ModelCleanable {
     newIdentity.identityId = rootPublicKey;
     newIdentity.type = Identity_Type.ECDSA;
 
+// If this is the first identity, set its analytics ID to the default one
+    if (((Globals?.accountPool?.value?.length ?? 0) == 0) &&
+        ((Globals?.appState?.analyticsKey?.length ?? 0) > 0))
+      newIdentity.analyticsID = Globals.appState.analyticsKey;
+    else
+      newIdentity.analyticsID = generateAnalyticsKey();
+
     dvote.Key k = dvote.Key();
     k.type = Key_Type.SECP256K1;
     k.encryptedMnemonic = encryptedMenmonic;
@@ -489,4 +497,8 @@ class AccountModel implements ModelRefreshable, ModelCleanable {
 
     return AccountModel.fromIdentity(newIdentity);
   }
+}
+
+String generateAnalyticsKey() {
+  return randomString();
 }