+++ lib/views/startup-page.dart
@@ -1,7 +1,9 @@
 import 'dart:developer';
 import 'package:dvote_common/widgets/loading-spinner.dart';
 import 'package:flutter/material.dart';
+import 'package:package_info/package_info.dart';
 import 'package:vocdoni/app-config.dart';
+import 'package:vocdoni/lib/analtyics.dart';
 import 'package:vocdoni/lib/i18n.dart';
 import 'package:vocdoni/lib/notifications.dart';
 import 'package:vocdoni/lib/startup.dart';
@@ -27,8 +29,6 @@ class _StartupPageState extends State<StartupPage> {
   }
 
   Future<void> initApplication() {
-    Globals.analytics.init();
-
     if (!mounted) return Future.value();
     setState(() {
       loading = true;
@@ -36,8 +36,13 @@ class _StartupPageState extends State<StartupPage> {
     });
 
     return restorePersistence()
-        .then((_) => restoreDataPools()) // Depends on restorePersistence()
+        .then((_) => restoreDataPools())
+        // Depends on restorePersistence()
         .then((_) => Notifications.init())
+        .then((_) => AppConfig.setPackageInfo())
+        .then((_) => AppConfig.setDeviceInfo())
+        .then((_) => Globals.analytics.init())
+        .then((_) => Globals.analytics.trackEvent(Events.APP_START))
         .then((_) {
       showNextScreen();
 
@@ -51,6 +56,8 @@ class _StartupPageState extends State<StartupPage> {
         loading = false;
         error = getText(context, "error.couldNotReadInternalData");
       });
+      Globals.analytics.init();
+      Globals.analytics.trackError("AppStartupError: $err");
 
       // RETRY ITSELF
       Future.delayed(Duration(seconds: 10)).then((_) => initApplication());